Preparatory: Associate Programming Lang/Article
In order to search for articles by programming language, an association
is required.

This ensures that an article belongs to a programming language.

When the search includes a programming language, articles only that
belong to that programming language will be selected.

shoulda-matchers gem was added to provide much needed model
matchers that make the tests more fluent.