Feature: Visitor can login
In order for a visitor to gain access to certain defined privileges,
the visitor is required to login.

This implementation make it possible for a visitor to become an
authenticated one.

This commit also comprises of end to end testing of the visitor's
journey (i.e pre and post login) and unit tests.