Preparatory: Request for password reset
As part of the implementation to enable a visitor with an account to
request the reset of the password on their account.

This commit presents the visitor with a link, form and email message
with the aim of assisting the visitor to reset thier password.

This commit ends at the point where the user receives an email.

The next would include the actual update of the password.